Allow circular resource references in .tres files
This patch updates text ResourceFormatSaver/Loader to
* do not crash when dealing with circular resource references
* instead of breaking references it stores them in a way that
  allowes properly restoring references back on load.

To fix a crash when saving resources with circular references,
_find_resources() method needed to be updated to store resource
into resource_set right after check if resource in the set. That
prevents infinite loop (and eventual stack overflow) when
traversing fields of a resource (which could lead back to
previously seen resource, thus creating an infinite traversal
loop).

The problem with storing circular references is that original
implementation assumed that resources form a tree. Then it
created topological ordering of them where children always
appeared before their parent.

The idea behind new algorithm is that resource loops can be
broken at a point and one of the resources need to be forward
declared (in a resource file, and created upfront during load)
and its initialization can happen later.

This if there is a resource loop A -> B -> C -> B with A being
main resource, it would find during topological ordering that
reference C -> B introduces a loop. Normally, it would store
resources in an order C, B, A, but since construction of C
requires B, it would first forward declare B, then would do
normal resource declaration in the same order:

[gd_resource type="A" load_steps=4]
[sub_resource type="B" id=1]  # <- forward declaration, no properties
[sub_resource type="C" id=2]
[sub_resource type="B" id=1]  # subsequent occurance of the same subresource
[resource]

The idea is that subsequent declaration of a resource overwrites
(or updates) properties declared previously. Forward declarations
look just like regular resources without any properties set.

In case of A -> B -> C -> A loop, the loop will be broken on C -> A
reference, except for main resource does not need a forward declaration,
main tag ("gd_resource") already serves as one (it contains a type
and id is always 0). Later, main resource can be referenced as
SubResource(0).